+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This is the changelog for [Authress Login](readme.md).
 * Remove deprecated properties from previous version.
 * Increase time span for duplicate session checks to 50ms
 * Add automatic retries to network connection issues.
+* Add AuthUserId cookie available in all requests to replace the `user` cookie.
 
 ## 1.4 ##
 * Include the extensionClient as embeddable client for OAuth extension login.